third_party/blink/web_tests/http/tests/misc/bubble-drag-events-expected.txt

CONSOLE ERROR: Not allowed to load local resource: file:
This page tests the drag event bubbling with valid and invalid targets.

Drag here.
Tests that we don't bubble up the events if we drag something to an iframe that has an invalid source
Tests that we don't bubble up the events if we drag something to an iframe that has a valid source
Tests that we bubble up the events if we drag something to a non-iframe target
ondragenter
ondragover
ondrop
